+++ lib/src/exceptions.dart
@@ -1,0 +1,5 @@
+class ClosedQueueException implements Exception {
+  final String message;
+
+  ClosedQueueException(this.message);
+}